requirements.txt
pathlib
python-dotenv
google-generativeai
youtube_transcript_api
streamlit